+++ yt_queue/cli.py
@@ -29,6 +29,9 @@ def _add_refresh_sub_command(subparsers):
     parser_refresh = subparsers.add_parser('refresh',
         help="Refresh the playlist, updating the videos in the given file")
     parser_refresh.add_argument('file')
+    parser_refresh.add_argument('--only-if-older', dest='only_if_older',
+                                metavar='SINCE',
+                                help='Skip if refreshed more recently than <x><hours|mins|days>')
 
 def _add_info_sub_command(subparsers):
     parser_info = subparsers.add_parser('info',