+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## unreleased
+
+### Added
+
+- CLI: add `--only-if-older=<time>` to `refresh` sub command to skip if recently refresed
+- api: `only_if_older` option to refresh to skip refreshing if recently refreshed
+
 ## 0.4.0
 
 ### Added